(SIMP-9883) Generalize simpkv acceptance tests (#63)
Rewrote simpkv acceptance tests used by the file plugin
so that the test infrastructure can be used by any plugin
or any combination of plugins. Also, addressed test
deficiencies such as the absence of global key testing and
test infrastructure design flaws such as the confusing,
unnecessary coupling of test data and the simpkv_test module.

This test infrastructure is required for the ldap plugin.